Fix slugs
